Move post model tests out of Http tests.
